[BUGFIX] Updating page in node tree can lead to problems
When a page is being edited in the node tree and the url is changed
on the server side, the link in the tree is out of sync.

Additionally if the title of the current page is changed, the page
is not reloaded.

To prevent these situations the URL of the node in the tree is updated,
as well as the title and if it's the current page it's reloaded automatically.

Change-Id: Ie71bf4f90b51ff9dfa46ba993f4a4f52ee3e0aa4
Releases: master, 1.2, 1.1
Fixes: NEOS-1197