[BUGFIX] Moving node to a new parent and back makes it disappear
When moving a node to a new parent and back to it's original parent
again makes it disappear. This happens because the `NodeOperations`
move method checks if there already exists a node in that location
in any context. Since there is a shadow node in it's original place,
that returns true and thus a new name is generated to avoid a conflict.
Problem is that the shadow node and the node will then share the
identifier on the same level, which leads to the node disappearing.

Change-Id: I853c4885fd694e0e5f73ce315ea67ee1da492b9c
Releases: master, 1.2
Fixes: NEOS-1203