[TASK] Insert content without page reload when creating/pasting
When new content is created or existing content is pasting using cut or copy,
the page reload is replaced with directly inserting of the content. This is
done to improve user experience for editors when creating/pasting content since
it interrupts the flow, especially when working on long pages with lots of content.

Change-Id: I0fbcc47dedb643563c2e2af072c59038a94d33b8
Releases: master, 2.0
Resolves: NEOS-613