[BUGFIX] Allow ``PluginViews`` to encapsulate multiple actions
Currently the ``PluginViewImplementation`` only evaluates the default
controller/action pair that is configured in the ``controllerActions``
setting of the corresponding NodeType.
Additional controllers/actions are ignored so that it's not possible to
invoke them.

This is fixed by falling back to the default controller/action only if
no explicit action is requested.

Change-Id: Id7aa9c110280d23198673b5f90d5a4eefa98a003
Fixes: NEOS-1311
Releases: master, 2.0, 1.2