[BUGFIX] ContentCollectionRenderer should not taint Node context
The ContentCollection also renders removed nodes while in editing mode,
so publishing can pick up those nodes and give a correct count of changes
available for publishing.

Without this change removed and non removed nodes are fetched in one
expression, but that results in all visible nodes to have a "tainted"
context that permits seeing removed nodes. This results in removed nodes
being fetched if a content element fetches and renders sub nodes without
using the ContentCase renderer (as that adds a class that hides removed
elements).

This change fixes the problem by separating the retrieval of removed nodes
from that of visible nodes. That way only the removed nodes get the
tainted context but as they are not rendered visible that won't be
a problem.

Change-Id: I187e007d9a8b548265b1bd531d294cee6b63e088
Releases: master, 2.0