[BUGFIX] Too many missing translation logs
To allow fallbacks to translations configured on supertypes
of nodetypes we need to call the translate method twice which
would result in a 'Missing translation' warning even though the
fallback translation is found.

This change overwrites the default Ember.Logger behavior and adds
a mute flag to prevent unneeded logs.

Change-Id: Ice57aa749df953af2fc377a38d4e7aedb1daf240
Related: NEOS-437
Releases: master, 2.0