[BUGFIX] Selecting editable requires a second click
Fix regression caused by ``I6f04b5eed47193a5322514be67d179c593ed1ba8``
where editables are not selected on first click. This happens because
when focussing a new element the active editables are deselected, but
the newly focussed editable is focussed before the update of the
selected element and thus deselected for the newly selected element.

Introduce ``deselectEditables`` option for ``updateSelection()``,
to only try to deselect editables when it's desired.

Change-Id: I59446b9490b877136e3bebcfb2a74e23ff0a273e
Releases: 2.0, master