[BUGFIX] Avoid mismatching finalImageDimensions on load
Due to the bindings we have, it could happen that the provided
metadata for an image contain values for an ``ResizeAdjustment``
that are slightly different from the calculated sizes by using a
ratio generated from the crop values. This mismatch would lead to
a value difference and applicable changes for an image right after
selecting it.

The change circumvents that by setting the dimensions straight from
the provided metadata (again) after bindings have triggered.

Dimensions will still be updated when cropping so it could happen that
the crop area is modified in a way that you end up with the same crop
area and still have a modification afterwards due to the
calculations done. This is probably acceptable.

Change-Id: I47714b5323c5e871f8c6024ba9322942c0fd1ccc
Resolves: NEOS-1384
Releases: master, 2.0