[TASK] Make PHP highlighting work for snippets
This adds the "official" hack to turn on startinline for all php
codeblocks whether they're a snippet or a full file.

Sphinx and docutils don't support setting the startinline option for
pygments. That means that snippets of PHP code that don't have
"<?php ?>" don't get highlighted. The hack was documented here:
http://mbless.de/blog/2015/03/02/php-syntax-highlighting-in-sphinx.html
https://github.com/fabpot/sphinx-php

Change-Id: I8e8863baa1a3c50c704b6e0318b08b8ee9c037cb
Releases: master, 2.0, 1.2, 1.1, 1.0