[TASK] Add support for sphinx-autobuild
This adds support for sphinx-autobuild, which watches the Documentation
directory and automatically re-renders docs on change. This allows for
a livepreview while editing the docs.

To use it, install `sphinx-autobuild`_::

    pip install sphinx-autobuild

To use livepreview just run ``make livehtml`` instead of ``make html``.
Then, visit http://127.0.0.1:8000 to see the livereload version of the
docs. Note, this builds the docs in ``_build/livehtml`` instead of
``_build/html`` because the docs include some javascript to make
livereload work.

.. _sphinx_autobuild: https://github.com/GaretJax/sphinx-autobuild

Change-Id: Ib16a73c565bae5388c38cd6faa41c78684d66d42
Releases: master, 2.0, 1.2, 1.1, 1.0