[BUGFIX] Avoid side-effects when rendering content for node creation/updating
The ``Service\NodeController`` which is used to render nodes via
AJAX in backend operations breaks on inserting plugins as linking in
plugins is broken due to a wrong parent request.

This change moves rendering fully to the ``Frontend\NodeController``
so the parent request for plugins is always correct. It triggers a
redirect to the showAction with internal arguments needed to render
the node out of band.

Change-Id: I18f86eeaa9f47737506fdb289833c2fe982bcb79
Fixes: NEOS-1368
Releases: master, 2.0, 1.2, 1.1