[TASK] Rework documentation structure
Also include auto generated docs of more packages.

Structure of the Documentation:

* Getting Started
* Technical Principles
* User Guide
* Creating a Site
* Extending Neos
* Inside of Neos
* References
* Contribute
* How Tos
* Running Neos
* Appendixes

The configuration for the auto generated documentations was moved into
the settings of the Neos package and was extended to cover all packages
that are part of a Neos demo installation.

The existing parts were not altered yet. They were simply moved into the
new structure.

Change-Id: I87ca0bceb785e02efc66dfd339cbc0f8bc8d971e
Resolves: NEOS-1376
Releases: master, 2.0