[BUGFIX] Return backend user even if security context is uninitialized
If the security context has not yet been initialized the user service
would return NULL instead of initializing the security context and then
check if there was a backend user. Instead of only returning it if it's
already initialized, it checks if it's possible to initialize or not.

Change-Id: I310dc040463e465247a379700fd03fca9c01cc83
Releases: master, 2.0, 1.2, 1.1