[BUGFIX] Prevent client-side performance issue when refreshing editables
Refreshing an editable should take the same approach as enabling the
editing in `create.js`. There could be sever client-side performance
issues when calling `midgardEditable` on an element that has lots of
nested editables because that will end up in many RDFa entity
extractions (VIE load).

This changes implements an explicit loading of entities from RDFa from
the given element and gives the `midgardEditable` widget an explicit
model.

Change-Id: I2f6c3c4f50ecbf8259c62fd3859396291d4007f7
Releases: master, 2.0
Fixes: NEOS-1405