[TASK] Clean up Release Notes and ChangeLogs
The 1.0 release notes are merely links to the Changelogs.
The >=1.1 release notes could be useful even when they are for older
versions of the docs. So, this archives <1.0 and keeps >1.0.

This also adds a :reversed: flag to the TOC, so that the the release
notes can be listed in reverse chronological order.

This also updates the ChangeLogs so that they don't loose all of the
formatting fixes that were added in later branches. Now that this is the
only copy of those ChangeLogs, they need to be as usable as possible.

Change-Id: Iff4bee00f20ae46fcaf4e3f37d35fcd66dfc1d63
Releases: master, 2.0, 1.2, 1.1, 1.0