[BUGFIX] Discarding changes doesn't reset link inspector editor
When discard changes in the inspector, link editors aren't updated
correctly. If a value is selected and then discarded, the newly
selected value remains. If a value has already been selected and
that is removed, discarding that change will result in a the input
field appearing when there's already a value selected.

Change-Id: Ib5a86ecb4f112a26b6fe568f8845670d057513f1
Releases: master, 2.0, 1.2
Fixes: NEOS-135