[BUGFIX] Missing node type configuration causes JS load error
If the node type configuration cannot be found for the current document
node a fatal error is thrown in the navigate component. This change makes
sure to properly test if the configuration is available before using it,
to prevent such errors.

Change-Id: Ic0e1d630d05c0087209ab3205eb94466783930e8
Releases: master, 2.0
Fixes: NEOS-1478
(cherry picked from commit 4417789175c4476538d4055fc3929f2992288155)